Implemented permutation and unsqueeze views
Implemented PermutedDimsView, which represents a view into another Tensor with dimensions permuted by some arbitrary permutation. Along with this, the swapAxes, t, permuteDims and moveAxis methods were added, which all use this view.
Implemented UnsqueezeView, which adds in singleton dimensions, and the unsqueeze method, which allows use of this view.
Refactored how elements are accessed. Now, subclasses only need to override one function which maps external to internal indices. The base Tensor is now native to the Tensor class, and get/set will use it if data is null.
Added toArray function with generator and converter, to allow conversions to arrays of any type.
Added validateAxis method to simplify axis error checking.
Added a new overload to the apply method, which allows the applied function to operate on both the flattened index and the value